cli/requirements.txt
boto3>=1.16.14
future>=0.18.2
tabulate>=0.8.2,<=0.8.7
ipaddress>=1.0.22
enum34>=1.1.6
configparser>=3.5.0
PyYAML>=5.3.1
jinja2>=2.11.0
marshmallow
aws-cdk.core
aws-cdk.aws-ec2
